コードフォーマットを適用しました。
Update  issue 31 

コードフォーマットを適用しました。
・ ToDoリストポートレット



git-svn-id: http://aipo.googlecode.com/svn/trunk@103 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ portlets/todo/src/main/java/com/aimluck/eip/todo/ToDoSelectData.java
@@ -202,7 +202,8 @@ public class ToDoSelectData extends ALAbstractSelectData<EipTTodo> implements
       ToDoResultData rd = new ToDoResultData();
       rd.initField();
       rd.setTodoId(record.getTodoId().intValue());
-      rd.setCategoryId(record.getEipTTodoCategory().getCategoryId().longValue());
+      rd.setCategoryId((int) record.getEipTTodoCategory().getCategoryId()
+        .longValue());
 
       rd.setCategoryName(ALCommonUtils.compressString(record
         .getEipTTodoCategory().getCategoryName(), getStrLength()));
